+++ lib/src/models/helper/user_tag_model.dart
@@ -1,4 +1,5 @@
 import 'package:json_annotation/json_annotation.dart';
+import 'package:likeminds_feed/src/models/sdk/sdk_client_info_model.dart';
 
 part 'user_tag_model.g.dart';
 
@@ -9,6 +10,7 @@ class UserTag {
   final int? id;
   final bool? isGuest;
   final String? userUniqueId;
+  final SDKClientInfo? sdkClientInfo;
 
   UserTag({
     this.name,
@@ -17,6 +19,7 @@ class UserTag {
     this.id,
     this.isGuest,
     this.userUniqueId,
+    this.sdkClientInfo,
   });
 
   factory UserTag.fromEntity(UserTagEntity entity) {
@@ -27,6 +30,7 @@ class UserTag {
       id: entity.id,
       isGuest: entity.isGuest,
       userUniqueId: entity.userUniqueId,
+      sdkClientInfo: entity.sdkClientInfo,
     );
   }
 
@@ -38,6 +42,7 @@ class UserTag {
       id: id,
       isGuest: isGuest,
       userUniqueId: userUniqueId,
+      sdkClientInfo: sdkClientInfo,
     );
   }
 
@@ -62,6 +67,9 @@ class UserTagEntity {
   @JsonKey(name: 'user_unique_id')
   final String? userUniqueId;
 
+  @JsonKey(name: 'sdk_client_info')
+  final SDKClientInfo? sdkClientInfo;
+
   UserTagEntity({
     this.name,
     this.imageUrl,
@@ -69,6 +77,7 @@ class UserTagEntity {
     this.id,
     this.isGuest,
     this.userUniqueId,
+    this.sdkClientInfo,
   });
 
   factory UserTagEntity.fromJson(Map<String, dynamic> json) =>