+++ CHANGELOG.md
@@ -58,3 +58,12 @@ Adding Save post
 # 1.3.1
 
 Fixed tagging api response
+
+# 1.3.2
+
+Make LMSDKCallback optional
+
+# 1.3.3
+
+Fixed tagging api response
+Adding logout service